Put back timestamp into union of two 32bits integers
git-svn-id: https://svnsrv.desy.de/desy/EuXFEL/WP76/karabo/karaboFramework/trunk@9952 a23ab9f9-1ea2-4934-960d-990259f9ea8c

+++ src/karabo/util/Timer.hh
@@ -41,8 +41,8 @@ namespace karabo {
 
             struct {
 
-                uint64_t nsec;
-                uint64_t sec;
+                uint32_t nsec;
+                uint32_t sec;
             };
         } timestamp;
